cambios finales a la estructura general
