+++ INEGIpy/__indicadores/__init__.py
@@ -1,2 +1,2 @@
-from .indicador_general import IndicadorGeneral
+from .inegi_general import INEGI_General
 from .pib import PIB
